orchestrator/requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
boto3==1.37.23
python-multipart==0.0.20
python-dotenv==1.0.1
prefect==2.20.16
pydantic==2.9.2
PyYAML==6.0.2
docker==7.1.0
numpy==2.2.5
h5py==3.13.0
